Consolidate participant stage data triggers into new onParticipantStageUpdated trigger
