+++ functions/src/stages/chat.triggers.ts
@@ -83,77 +83,6 @@ export const onChatMessageCreated = onDocumentCreated(
   },
 );
 
-/** When chat participant answer is updated, check if all participants
- * are ready to move to next discussion. */
-export const updateCurrentChatDiscussionId = onDocumentWritten(
-  {
-    document:
-      'experiments/{experimentId}/participants/{participantId}/stageData/{stageId}',
-    timeoutSeconds: 60,
-  },
-  async (event) => {
-    const data = event.data.after.data() as StageParticipantAnswer | undefined;
-
-    const stageDocument = app
-      .firestore()
-      .collection('experiments')
-      .doc(event.params.experimentId)
-      .collection('stages')
-      .doc(event.params.stageId);
-    const stage = (await stageDocument.get()).data() as StageConfig;
-    if (stage.kind !== StageKind.CHAT) return;
-
-    // Define participant document
-    const participantDocument = app
-      .firestore()
-      .collection('experiments')
-      .doc(event.params.experimentId)
-      .collection('participants')
-      .doc(event.params.participantId);
-
-    const participant = (
-      await participantDocument.get()
-    ).data() as ParticipantProfileExtended;
-
-    // Define public stage document reference
-    const publicDocument = app
-      .firestore()
-      .collection('experiments')
-      .doc(event.params.experimentId)
-      .collection('cohorts')
-      .doc(participant.currentCohortId)
-      .collection('publicStageData')
-      .doc(event.params.stageId);
-
-    await app.firestore().runTransaction(async (transaction) => {
-      // Update public stage data
-      const publicStageData = (
-        await publicDocument.get()
-      ).data() as StagePublicData;
-      const discussionStatusMap = data.discussionTimestampMap;
-
-      for (const discussionId of Object.keys(discussionStatusMap)) {
-        if (!publicStageData.discussionTimestampMap[discussionId]) {
-          publicStageData.discussionTimestampMap[discussionId] = {};
-        }
-        publicStageData.discussionTimestampMap[discussionId][
-          participant.publicId
-        ] = discussionStatusMap[discussionId];
-      }
-
-      // Update current discussion ID if applicable
-      await updateCurrentDiscussionIndex(
-        event.params.experimentId,
-        participant.currentCohortId,
-        event.params.stageId,
-        publicStageData,
-      );
-
-      transaction.set(publicDocument, publicStageData);
-    });
-  },
-);
-
 /** When chat message is created, generate mediator agent response if relevant. */
 // TODO: Rename to createAgentMediatorMessage
 export const createAgentMessage = onDocumentCreated(
